refactored tests into separate files. cleaned up vestigial testfile. cleaned up vestigial reducer and its tests. documentation tweaks accross the repo.
